requirements.txt
google-generativeai
llama-index-core
llama-index-llms-gemini
llama-index-embeddings-huggingface
sentence-transformers
torch
torchvision
pdfplumber
pypdf
pdf2image
pytesseract
python-pptx
python-dotenv
streamlit
sympy
numpy
matplotlib
pint
requests
